requirements.txt
colorama==0.4.6
et_xmlfile==2.0.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.2
qrcode==7.4.2
six==1.16.0
typing_extensions==4.12.0
tzdata==2024.2
